Create Weibull
The Weibull parametric model is a widely used survival analysis technique that assumes the time-to-event (survival time) follows a Weibull distribution. The Weibull model is flexible, as it can model both constant, increasing, or decreasing hazard rates, depending on the value of the shape parameter (β). This makes it useful for a variety of survival data types.